+++ parameters_unpacking_variables.py
@@ -1,0 +1,8 @@
+from sys import argv
+# read the WYSS section for how to run this
+script, first, second, third = argv
+
+print("The script is called:", script)
+print("Your first variable is:", first)
+print("Your second variable is:", second)
+print("Your third variable is:", third)
