+++ prompting_people.py
@@ -1,0 +1,5 @@
+age = input("How old are you? ")
+height = input("How tall are you? ")
+weight = input("How much do you weigh? ")
+
+print(f"So, you're {age} old, {height} tall and {weight} heavy.")
